refactor: Remove experimental TUI feature
The TUI was an experiment to explore interactive interfaces. However,
tagging and file management is better suited for a GUI which supports
thumbnails. Developing a TUI will only bring pain and suffering and
ultimately result in software that is categorically worse than a proper
GUI.

The complete TUI feature has been archived on the 'tui-archive' branch
for future reference.